API 100% funcional, falta passar a vassoura

+++ BadgeSpace/Extensions/AlunosController.cs
@@ -2,6 +2,7 @@
 using BadgeSpace.Data.Migrations;
 using BadgeSpace.Models;
 using BadgeSpace.Models.Enums;
+using BadgeSpace.Utils.Security;
 using Humanizer;
 using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Mvc;
@@ -9,6 +10,7 @@ using Microsoft.EntityFrameworkCore;
 
 namespace BadgeSpace.Extensions
 {
+    // falta criar uma função pra reduzir a repetição do código
     [Controller]
     [Route("api/[controller]")]
     public class AlunosController : ControllerBase
@@ -17,7 +19,7 @@ namespace BadgeSpace.Extensions
 
         public AlunosController(ApplicationDbContext context) => _context = context;
 
-        [HttpGet("email={email}/{CPF?}/{ID?}")]
+        [HttpGet("listar/{email}/{CPF?}/{ID?}")]
         [Authorize(Roles = nameof(Roles.EMPRESS))]
         public async Task<IActionResult> Listar(string email, string? CPF, int? ID)
         {
@@ -29,7 +31,6 @@ namespace BadgeSpace.Extensions
                 return Ok(new { student.Id, student.NomeAluno, student.AlunoCPF, student.Curso });
             }
                 
-
             if (CPF != null && CPF != "")
             {
                 var students = _context.Students
@@ -49,46 +50,94 @@ namespace BadgeSpace.Extensions
             return Ok(dados);
         }
 
-        [HttpGet("{CPF}")]
+        [HttpGet("{CPF}/{empresa?}/{curso?}")]
         [Authorize(Roles = nameof(Roles.STUDENT))]
-        public IActionResult ListarCertificados(string? CPF)
+        public async Task<IActionResult> ListarCertificados(string CPF, string? empresa, string? curso)
         {
-            if (CPF != null)
+            var confirmC = await _context.Users.FirstOrDefaultAsync(c => c.CPF_CNPJ == CPF);
+            if (confirmC == null)
+                return NotFound(new { message = "Inválido." });
+            if (empresa != null)
             {
-                var student = _context.Students.Where(x => x.AlunoCPF == CPF);
-                if (student == null) return NotFound(new { message = "Aluno Inválido." });
-
+                var confirmE = await _context.Users.FirstOrDefaultAsync(c => c.Email == empresa);
+                if (confirmE == null)
+                    return NotFound(new { message = "Inválido." });
+            }
+                
+            if (curso != null && curso != "")
+            {
+                var student = _context.Students.Where(c => c.EmpresaId == empresa && c.AlunoCPF == CPF && c.Curso == curso);
                 return Ok(student);
             }
 
-            return BadRequest();
+            if (empresa != null && empresa != "")
+            {
+                var student = _context.Students.Where(c => c.EmpresaId == empresa && c.AlunoCPF == CPF);
+                return Ok(student);
+            }
+
+            var Student = _context.Students.Where(c => c.AlunoCPF == CPF);
+
+            return Ok(Student);
         }
 
-        [HttpPost]
+        [HttpPost("{empresa}/cadastrar={CPF}")]
         [Authorize(Roles = nameof(Roles.EMPRESS))]
-        public async Task<IActionResult> Cadastrar([FromBody] StudentModel Student)
+        public async Task<IActionResult> Cadastrar(string CPF, string empresa, [FromBody] StudentModel Student)
         {
+            var confirmE = await _context.Users.FirstOrDefaultAsync(c => c.Email == empresa);
+            var confirmC = await _context.Users.FirstOrDefaultAsync(c => c.CPF_CNPJ == CPF);
+
+            if (confirmC == null || confirmE == null)
+                return NotFound(new { message = "Inválido." });
+
+            Student.AlunoCPF = CPF;
+            Student.EmpresaId = empresa;
+
             _context.Students.Add(Student);
             await _context.SaveChangesAsync();
 
             return Created("Aluno", Student);
         }
 
-        [HttpPut]
+        [HttpPut("{empresa}/editar={CPF}/{ID?}")]
         [Authorize(Roles = nameof(Roles.EMPRESS))]
-        public async Task<IActionResult> Editar([FromBody] StudentModel Student)
+        public async Task<IActionResult> Editar(string CPF, string empresa, int? ID, [FromBody] StudentModel? Student)
         {
+            var confirmE = await _context.Users.FirstOrDefaultAsync(c => c.Email == empresa);
+            var confirmC = await _context.Users.FirstOrDefaultAsync(c => c.CPF_CNPJ == CPF);
+
+            if (confirmC == null || confirmE == null)
+                return NotFound(new { message = "Inválido." });
+
+            if (Student!.EmpresaId != empresa)
+                return NotFound(new { message = "Aluno Inválido." });
+
+            if (Student!.AlunoCPF != CPF)
+                return NotFound(new { message = "Aluno Inválido." });
+
+            if (ID.HasValue && ID != 0)
+                if (Student.EmpresaId != empresa)
+                    return NotFound(new { message = "Aluno Inválido." });
+
             _context.Students.Update(Student);
             await _context.SaveChangesAsync();
             return Ok(Student);
         }
 
-        [HttpDelete("deletar={CPF}/{ID?}")]
+        [HttpDelete("{empresa}/deletar={CPF}/{ID?}")]
         [Authorize(Roles = nameof(Roles.EMPRESS))]
-        public async Task<IActionResult> Remover(string CPF, int? ID)
+        public async Task<IActionResult> Remover(string empresa, string CPF, int? ID)
         {
+            var confirmE = await _context.Users.FirstOrDefaultAsync(c => c.Email == empresa);
+            var confirmC = await _context.Users.FirstOrDefaultAsync(c => c.CPF_CNPJ == CPF);
+
+            if (confirmC == null || confirmE == null)
+                return NotFound(new { message = "Inválido." });
+
             var student = ID.HasValue && ID != 0 
-                ? await _context.Students.FirstOrDefaultAsync(c => c.Id == ID) : await filtrar(CPF);
+                ? await _context.Students.FirstOrDefaultAsync(c => c.Id == ID) 
+                : await _context.Students.FirstOrDefaultAsync(c => c.AlunoCPF == CPF && c.EmpresaId == empresa);
             if (student == null) return NotFound(new { message = "Aluno Inválido." });
 
             _context.Students.Remove(student);
